Initial Model using T5-small
This model is based on the documentation of the T5 on Hugging Face.
It reads all assembly instructions and their corresponding C code as input and target for the seq-to-seq transformer. Calculates the loss with the current weights, and uses gradient descent to adjust them.

With moderate little training we already see first results. The transformer is able to predict the functions head more or less. The output is to small.
Please note we use T5-small for testing purposes. This transformer uses at most 512 Tokens which is not sufficient with respect to the sequence lengths of the assemblies we work with. Using T5-large is recommended.