Scraper.py
Can now handle multi-source-files in Github repositories. Scrapes repositories (with keywords) while maintaining the repo structure and tries to compile source files, by recursively finding missing headers.
Generates two folders: One with the source and header files, and one with the compiled object files.
Once one source file is possible to be compiled, the Scraper prefixes a gcc command with include arguments that helps to easily compile the file in the future again.
An exact usage is documented in the script itself.